Move attribute utils to its own util class
- Our singular test failed because we couldn't access bukkit's registry during testing (which we access statically for attribute stuff). So, we moved the attribute stuff to its own AttributeUtil class.
